requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pyclipper==1.4.0
Pygments==2.19.2
pyparsing==3.3.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.1
setuptools==82.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2026.2.20
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.15
ultralytics-thop==2.0.18
urllib3==2.6.3
